+++ nodedb-fts/src/backend/traits.rs
@@ -8,7 +8,7 @@ use crate::posting::Posting;
 /// in-memory HashMap. All scoring, BMW, compression, and analysis logic
 /// works identically over any backend.
 ///
-/// Every tenant-partitioned method takes `tid: u32` as a first-class
+/// Every tenant-partitioned method takes `tid: u64` as a first-class
 /// parameter. Backends are required to isolate tenants structurally —
 /// no tenant boundary may depend on lexical-prefix ordering of a
 /// composed string key.
@@ -25,7 +25,7 @@ pub trait FtsBackend {
     /// Read the posting list for a term in a collection.
     fn read_postings(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         term: &str,
     ) -> Result<Vec<Posting>, Self::Error>;
@@ -33,19 +33,19 @@ pub trait FtsBackend {
     /// Write/replace the posting list for a term in a collection.
     fn write_postings(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         term: &str,
         postings: &[Posting],
     ) -> Result<(), Self::Error>;
 
     /// Remove a term's posting list entirely.
-    fn remove_postings(&self, tid: u32, collection: &str, term: &str) -> Result<(), Self::Error>;
+    fn remove_postings(&self, tid: u64, collection: &str, term: &str) -> Result<(), Self::Error>;
 
     /// Read the document length (token count) for a document.
     fn read_doc_length(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         doc_id: Surrogate,
     ) -> Result<Option<u32>, Self::Error>;
@@ -53,7 +53,7 @@ pub trait FtsBackend {
     /// Write/replace the document length for a document.
     fn write_doc_length(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         doc_id: Surrogate,
         length: u32,
@@ -62,33 +62,33 @@ pub trait FtsBackend {
     /// Remove a document's length entry.
     fn remove_doc_length(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         doc_id: Surrogate,
     ) -> Result<(), Self::Error>;
 
     /// Get all term names in a collection (for fuzzy matching).
-    fn collection_terms(&self, tid: u32, collection: &str) -> Result<Vec<String>, Self::Error>;
+    fn collection_terms(&self, tid: u64, collection: &str) -> Result<Vec<String>, Self::Error>;
 
     /// Get total document count and sum of all document lengths for a collection.
     /// Returns `(doc_count, total_token_sum)`.
     ///
     /// Implementations should maintain these incrementally for O(1) lookup.
-    fn collection_stats(&self, tid: u32, collection: &str) -> Result<(u32, u64), Self::Error>;
+    fn collection_stats(&self, tid: u64, collection: &str) -> Result<(u32, u64), Self::Error>;
 
     /// Increment collection stats after indexing a document.
     /// `doc_len` is the number of tokens in the newly indexed document.
-    fn increment_stats(&self, tid: u32, collection: &str, doc_len: u32) -> Result<(), Self::Error>;
+    fn increment_stats(&self, tid: u64, collection: &str, doc_len: u32) -> Result<(), Self::Error>;
 
     /// Decrement collection stats after removing a document.
     /// `doc_len` is the token count of the removed document.
-    fn decrement_stats(&self, tid: u32, collection: &str, doc_len: u32) -> Result<(), Self::Error>;
+    fn decrement_stats(&self, tid: u64, collection: &str, doc_len: u32) -> Result<(), Self::Error>;
 
     /// Read a metadata blob by sub-key (e.g., "docmap", "fieldnorms",
     /// "analyzer", "language").
     fn read_meta(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         subkey: &str,
     ) -> Result<Option<Vec<u8>>, Self::Error>;
@@ -96,7 +96,7 @@ pub trait FtsBackend {
     /// Write a metadata blob by sub-key.
     fn write_meta(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         subkey: &str,
         value: &[u8],
@@ -106,7 +106,7 @@ pub trait FtsBackend {
     /// identifier (e.g., `"L{level}:{id:016x}"`).
     fn write_segment(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         segment_id: &str,
         data: &[u8],
@@ -115,28 +115,28 @@ pub trait FtsBackend {
     /// Read a segment blob. Returns None if not found.
     fn read_segment(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         segment_id: &str,
     ) -> Result<Option<Vec<u8>>, Self::Error>;
 
     /// List all segment ids for a collection.
-    fn list_segments(&self, tid: u32, collection: &str) -> Result<Vec<String>, Self::Error>;
+    fn list_segments(&self, tid: u64, collection: &str) -> Result<Vec<String>, Self::Error>;
 
     /// Remove a segment blob.
     fn remove_segment(
         &self,
-        tid: u32,
+        tid: u64,
         collection: &str,
         segment_id: &str,
     ) -> Result<(), Self::Error>;
 
     /// Remove all entries for a collection. Returns count of removed entries.
-    fn purge_collection(&self, tid: u32, collection: &str) -> Result<usize, Self::Error>;
+    fn purge_collection(&self, tid: u64, collection: &str) -> Result<usize, Self::Error>;
 
     /// Remove all entries for a tenant across every collection. Returns
     /// count of removed entries. Implementations MUST use a structural
     /// drop (e.g., tuple range `(tid, ..)..(tid+1, ..)`) rather than a
     /// lexical-prefix scan.
-    fn purge_tenant(&self, tid: u32) -> Result<usize, Self::Error>;
+    fn purge_tenant(&self, tid: u64) -> Result<usize, Self::Error>;
 }
